seed working, modified one of the models, package.json name and db.js
