+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "boilermaker",
+  "name": "covid-store",
   "version": "2.0.0",
   "description": "Some boilerplate code to get you started - get shakin'!",
   "engines": {
