+++ server/db/db.js
@@ -1,8 +1,9 @@
 const Sequelize = require('sequelize')
 const pkg = require('../../package.json')
+const databaseName = pkg.name + (process.env.NODE_ENV === 'test' ? '-test' : '')
 
 const db = new Sequelize(
-  process.env.DATABASE_URL || `postgres://localhost:5432/covid-store`,
+  process.env.DATABASE_URL || `postgres://localhost:5432/${databaseName}`,
   {
     logging: false
   }
